SilverStripe 4 upgrade
