+++ _config.php
@@ -1,2 +1,1 @@
 <?php
-
