--- _config/extensions.yml
@@ -1,3 +1,0 @@
-SiteConfig:
-  extensions:
-    - 'TypeformSiteConfigExtension'
