+++ composer.json
@@ -1,15 +1,19 @@
 {
     "name": "dnadesign/silverstripe-typeform",
     "description": "Module to provide integration with Typeform. Downloads submissions and allows them to be managed within SilverStripe.",
-    "type": "silverstripe-module",
-    "keywords": ["silverstripe", "typeform"],
-	"license": "BSD-3-Clause",
+    "type": "silverstripe-vendormodule",
+    "keywords": [
+        "silverstripe",
+        "typeform"
+    ],
+    "license": "BSD-3-Clause",
     "authors": [{
         "name": "Will Rossiter",
         "email": "will.rossiter@dna.co.nz"
     }],
     "require": {
-        "silverstripe/framework": "~3.1"
+        "silverstripe/cms": "^4.4",
+        "silverstripe/admin": "^1.3"
     },
     "extra": {
         "installer-name": "typeform"
